+++ packages/zostso/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe z/OS TSO SDK package will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Add a new optional boolean field `useLegacyAPI` to `IIssueTsoCmdOpts`. For more details, see [the CLI changelog](../cli/CHANGELOG.md) [#2738](https://github.com/zowe/zowe-cli/issues/2738)
+
 ## `8.27.4`
 
 - BugFix: Updated minimum supported version of Node from 18 to 20. Added Node 24 support. [#2616](https://github.com/zowe/zowe-cli/pull/2616)