+++ packages/zostso/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Zowe z/OS TSO SDK package will be documented in this file.
 
-## Recent Changes
+## `8.33.0`
 
 - Enhancement: Added a new optional boolean field `useLegacyApi` to the `IIssueTsoCmdOpts` interface used to call the `IssueTso.issueTsoCmd` method. For more details, see [the CLI changelog](../cli/CHANGELOG.md). [#2738](https://github.com/zowe/zowe-cli/issues/2738)